requirements.txt
# ================================
# 青鸾向导 (QL Guide) - 项目依赖
# ================================

# -----------------
# Web框架和服务器
# -----------------
Flask==2.3.3
python-dotenv==1.0.0

# -----------------
# AI和语言模型
# -----------------
langchain==0.1.20
langchain-openai==0.1.8
langchain-core==0.1.52
langchain-community==0.0.38

# -----------------
# MCP协议支持
# -----------------
mcp==1.0.0
langchain-mcp-adapters==0.1.0

# -----------------
# HTTP客户端和网络
# -----------------
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.5

# -----------------
# 数据库和缓存
# -----------------
redis==5.0.1

# -----------------
# 数据处理和验证
# -----------------
pydantic==2.11.7

# -----------------
# 文档生成和处理
# -----------------
reportlab==4.4.2
markdown==3.8.2
pdfkit==1.0.0
weasyprint==62.3

# -----------------
# 搜索工具
# -----------------
duckduckgo-search==6.1.7

# -----------------
# 图形界面 (可选)
# -----------------
# tkinter 是Python内置模块，无需安装

# -----------------
# 开发和调试工具
# -----------------
colorama==0.4.6

# -----------------
# 异步支持
# -----------------
# asyncio 是Python内置模块，无需安装

# -----------------
# 时间和日期处理
# -----------------
# datetime 是Python内置模块，无需安装

# -----------------
# 安全相关
# -----------------
cryptography==42.0.8

# -----------------
# 其他实用工具
# -----------------
python-multipart==0.0.9
typing-extensions==4.12.2